[core][desktop]: Otimizar; Usar JavaFX; Adicionar testes

+++ core/src/main/kotlin/br/tiagohm/nestalgia/core/Register.kt
@@ -1,5 +1,6 @@
 package br.tiagohm.nestalgia.core
 
 interface Register {
-    val address: UShort
+
+    val address: Int
 }
